Version 2.5.1

+++ Automator/CreateEquations/Automator_CreateEquations.h
@@ -3,7 +3,7 @@
 //  Automator_CreateEquations
 //
 //  Created by Pierre Chatelier on 24/09/08.
-//  Copyright 2005, 2006, 2007, 2008, 2009, 2010, 2011 Pierre Chatelier. All rights reserved.
+//  Copyright 2005-2013 Pierre Chatelier. All rights reserved.
 //
 
 #import <Cocoa/Cocoa.h>
@@ -32,7 +32,7 @@
 
   ExportFormatOptionsPanes* generalExportFormatOptionsPanes;
   BOOL latexitPreferencesAvailable;
-  unsigned int uniqueId;
+  NSUInteger uniqueId;
   BOOL fromArchive;
 }
